init scan summary
